app/requirements.txt
streamlit==1.25.0
plotly==5.3.1
pandas==1.5.1
matplotlib==3.4.3
Pillow==9.3.0